% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdl-package.R
\docType{package}
\name{gdl-package}
\alias{gdl}
\alias{gdl-package}
\title{gdl: Graph Package Download Counts}
\description{
Plot new and cumulative download counts for an R package over time, combining CRAN download data (via the packageRank package) with optional GitHub Releases asset-download counts. Supports configurable time-unit aggregation and start-date filtering, and exposes the data layer via a custom S3 class with an autoplot() method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/gdl}
  \item \url{https://d-morrison.github.io/gdl/}
  \item Report bugs at \url{https://github.com/d-morrison/gdl/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu}

Authors:
\itemize{
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu}
}

}
\keyword{internal}
